doctor(finance): split Service in two and narrow the public contract
Service did the purchase-doc pipeline and the creditor bindings in one class.
They share no state and no invariant — one is a nightly full-pull with attribution
and an unmatched queue, the other a member-to-account link with a three-way
concurrency story — and the only thing joining them was that both talk to Holded.

  HoldedDocService  <- repo, client, IBudgetServiceRead, clock, logger
  CreditorService   <- repo, client, IHoldedService, clock, cache, logger

Each is registered once and exposed twice: a public contract in
Humans.Finance.Contracts carrying only what other sections call, and a wider
internal interface in Humans.Finance that only FinanceController resolves.
Provisioning, the unmatched queue, bind and unbind stop crossing the assembly
boundary; IHoldedFinanceService is gone.

Behaviour is unchanged — every method body moved verbatim. The tests split the
same way, and the two architecture lists that named the old class by string
(GDPR contributors, the IMemoryCache allowlist) now name CreditorService.

One wart survives the split: GetDocSyncInfoAsync is a doc-side method that reads
the creditor binding count, so /Holded can show it without building the full
creditor view. Both services share the section's one repository, so it stays
legal, but it is the seam's only leak.
