+++ src/Sections/Humans.Finance.Contracts/ICreditorService.cs
@@ -2,22 +2,13 @@ using Humans.Application.Interfaces;
 
 namespace Humans.Finance.Contracts;
 
-public interface IHoldedFinanceService : IApplicationService
+/// <summary>
+/// What other sections may ask of Finance's creditor bindings. Binding and unbinding are admin
+/// operations and are deliberately absent — they live on the section's internal
+/// <c>ICreditorAdminService</c>, which only <c>FinanceController</c> sees.
+/// </summary>
+public interface ICreditorService : IApplicationService
 {
-    Task<HoldedProvisioningPlan> GetProvisioningPlanAsync(int blockStart, CancellationToken ct = default);
-    Task<int> ProvisionAsync(int blockStart, bool addAll, CancellationToken ct = default);
-    Task<HoldedSyncResult> SyncAsync(CancellationToken ct = default);
-
-    /// <summary>State of the last purchase-doc sync. Read-only; lazy-creates the state row.</summary>
-    Task<HoldedDocSyncInfo> GetDocSyncInfoAsync(CancellationToken ct = default);
-    Task<IReadOnlyList<HoldedActualRow>> GetActualsForYearAsync(int calendarYear, CancellationToken ct = default);
-    Task<IReadOnlyList<HoldedUnmatchedRow>> GetUnmatchedAsync(CancellationToken ct = default);
-
-    /// <summary>The active Holded expense-account id mapped to this budget category
-    /// (holded_category_map), or null when no active mapping exists. Used to book a purchase
-    /// line's `items[].account` directly at doc creation.</summary>
-    Task<string?> GetHoldedAccountIdForCategoryAsync(Guid budgetCategoryId, CancellationToken ct = default);
-
     /// <summary>Derives cached creditor status (balance, owed, payments) for a member's 400000xx account
     /// from the cached daybook lines. Returns null when no lines are cached for the account.</summary>
     Task<HoldedCreditorStatus?> GetCreditorStatusAsync(
@@ -32,18 +23,13 @@ public interface IHoldedFinanceService : IApplicationService
     /// never resolved — neither our one-shot push resolution nor Holded's live contact list has a
     /// number for the contact — so there is no account row to place them on. There is no automatic
     /// retry (nobodies-collective/Humans#972), so returning them here is what keeps them visible on
-    /// /Finance/Creditors, where an admin binds them with <see cref="SetCreditorContactAsync"/>.
+    /// /Finance/Creditors, where an admin binds them.
     /// </returns>
     Task<(IReadOnlyList<HoldedCreditorAccountRow> Accounts, IReadOnlyList<CreditorContactBinding> Unresolved)>
         ListCreditorAccountsAsync(CancellationToken ct = default);
 
     /// <summary>The member's creditor-account binding, if any.</summary>
     Task<CreditorContactBinding?> GetCreditorContactByUserAsync(Guid userId, CancellationToken ct = default);
-
-    /// <summary>Manually binds a member to an existing Holded creditor account (by 400000xx number).
-    /// Resolves the Holded contact id. Fails — writing nothing — when the account is already bound to a
-    /// different member, or when no Holded contact carries that supplier-account number.</summary>
-    Task<CreditorBindResult> SetCreditorContactAsync(Guid userId, int supplierAccountNum, CancellationToken ct = default);
 
     /// <summary>Per-account statement: balance + itemized journal lines over the last ~year. Null if unknown.</summary>
     Task<HoldedCreditorLedger?> GetCreditorLedgerAsync(int supplierAccountNum, CancellationToken ct = default);
@@ -58,11 +44,4 @@ public interface IHoldedFinanceService : IApplicationService
 
     /// <summary>Records the resolved 400000xx number on the member's binding (once the payable exists).</summary>
     Task SetCreditorAccountNumAsync(Guid userId, int supplierAccountNum, CancellationToken ct = default);
-
-    /// <summary>Clears the member's creditor binding outright — the admin remedy for a wrong bind and
-    /// for the duplicate the automatic write paths record rather than refuse. Removes the whole row,
-    /// not just the 400000xx: a binding stripped of its number still carries the other member's Holded
-    /// contact id, which merges their payables just as thoroughly. The member's next expense push
-    /// re-resolves the contact from scratch. Returns false when there was nothing bound.</summary>
-    Task<bool> ClearCreditorContactAsync(Guid userId, CancellationToken ct = default);
 }
