+++ src/Sections/Humans.Finance/Section.cs
@@ -18,7 +18,7 @@ namespace Humans.Finance;
 /// Holded section, which registers it; Finance consumes it through
 /// <c>Humans.Holded.Contracts</c>. <c>HoldedSyncJob</c> is not registered here either, and is
 /// not this section's: it is a shim over Holded's <c>IHoldedNightlySync</c>, which calls this
-/// section's <c>IHoldedFinanceService.SyncAsync</c> first, and it lives in
+/// section's <c>IHoldedDocService.SyncAsync</c> first, and it lives in
 /// <c>Humans.Holded/Jobs/</c> since the HUM0034 Jobs carve-out.
 /// </remarks>
 public sealed class Section : ISection
@@ -28,9 +28,17 @@ public sealed class Section : ISection
         services.AddSectionDbContext<FinanceDbContext>(sentinelTable: "holded_expense_docs");
 
         services.AddScoped<IHoldedRepository, Repository>();
-        services.AddScoped<Service>();
-        services.AddScoped<IHoldedFinanceService>(sp => sp.GetRequiredService<Service>());
+
+        // Each service is registered once and exposed twice: the public contract other sections
+        // resolve, and the wider internal one FinanceController resolves.
+        services.AddScoped<HoldedDocService>();
+        services.AddScoped<IHoldedDocService>(sp => sp.GetRequiredService<HoldedDocService>());
+        services.AddScoped<IHoldedDocAdminService>(sp => sp.GetRequiredService<HoldedDocService>());
+
+        services.AddScoped<CreditorService>();
+        services.AddScoped<ICreditorService>(sp => sp.GetRequiredService<CreditorService>());
+        services.AddScoped<ICreditorAdminService>(sp => sp.GetRequiredService<CreditorService>());
         // Owns the user-scoped holded_creditor_contacts table → GDPR export contributor (design-rules §8a).
-        services.AddScoped<IUserDataContributor>(sp => sp.GetRequiredService<Service>());
+        services.AddScoped<IUserDataContributor>(sp => sp.GetRequiredService<CreditorService>());
     }
 }
